requirements.txt
beautifulsoup4==4.12.2
langchain==0.0.186
moviepy==1.0.3
nltk==3.7
numpy==1.23.5
omegaconf==2.2.3
openai_whisper==20230314
optimum==1.6.1
pandas==1.5.3
Pillow==9.5.0
Pillow==9.5.0
plotly==5.10.0
plotly_express==0.4.1
pyannote.audio==1.1.2
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==1.5.2
pydub==0.25.1
pytube==12.1.0
pyvis==0.3.2
PyYAML==5.4.1
PyYAML==6.0
ray==2.3.1
sentence_transformers==2.2.2
st_annotated_text==3.0.0
streamlit==1.16.0
streamlit_authenticator==0.2.1
torch==1.13.1
torchaudio==0.13.1
transformers==4.28.1
validators==0.18.2
vimeo_downloader==0.4.1
wget==3.2
wikipedia==1.4.0
en_core_web_lg @ https://huggingface.co/spacy/en_core_web_lg/resolve/main/en_core_web_lg-any-py3-none-any.whl
faster-whisper @ https://github.com/guillaumekln/faster-whisper/archive/a4f1cc8f11433e454c3934442b5e1a4ed5e865c3.tar.gz
